Method and Class Separation Groundwork
Gone right back to basics laying down the groundwork for having fully separate files for each method and class so that the main method is easy to approach and understand and not stacked with GUI commands alongside running the overall groundwork. Due to all the files needing to call upon variables which lie in other files. Specifically, the GUI.java  taking the generated password variable and displaying it as a JLabel, as such, you can see me learning and discerning just how I go about this.